Move formatting to new struct ChordChart
