+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod chord;
+pub mod chord_chart;
 pub mod chord_type;
 pub mod fret_pattern;
 pub mod interval;
@@ -9,6 +10,7 @@ pub mod tuning;
 pub mod voicing;
 
 pub use chord::Chord;
+pub use chord_chart::ChordChart;
 pub use chord_type::ChordType;
 pub use fret_pattern::FretPattern;
 pub use interval::Interval;
